+++ client/src/types/IContentWithLanguage.ts
@@ -4,6 +4,10 @@ export interface INoteContent {
   _content_html?: string
   equivalent?: Array<string>
   notation?: string
+  classifications?: Array<string>
+  languages?: Array<string>
+  label?: string
+  labelLanguages?: Array<string>
 }
 
 export interface IContentWithLanguage {